Fix type error in `interest.ts`
