+++ src/modules/mintNFT/nft.d.ts
@@ -1,0 +1,4 @@
+declare interface Window {
+  ethereum: any;
+  contract: any;
+}
